requirements.txt
tensorflow-gpu==1.11.0
numpy==1.15.4